+++ src/DMLQueryBuilder.php
@@ -29,7 +29,7 @@ use function substr;
 final class DMLQueryBuilder extends AbstractDMLQueryBuilder
 {
     /** @throws NotSupportedException */
-    public function insertWithReturningPks(string $table, array|QueryInterface $columns, array &$params = []): string
+    public function insertReturningPks(string $table, array|QueryInterface $columns, array &$params = []): string
     {
         throw new NotSupportedException(__METHOD__ . ' is not supported by MySQL.');
     }