+++ CHANGELOG.md
@@ -41,6 +41,7 @@
 - New #393: Use `DateTimeColumn` class for datetime column types (@Tigrov)
 - New #394, #395, #398: Implement `Command::upsertReturning()` method (@Tigrov)
 - Enh #394, #395: Refactor `Command::insertWithReturningPks()` method (@Tigrov)
+- Chg #399: Rename `insertWithReturningPks()` to `insertReturningPks()` in `Command` and `DMLQueryBuilder` classes (@Tigrov)
 
 ## 1.2.0 March 21, 2024